chore: 完成英文config的配置
