+++ docs/src/en/guide/introduction/thin.md
@@ -26,7 +26,7 @@ If you don’t need demo code, you can simply delete the `playground` folder
 
 If you don’t need documentation, you can delete the `docs` folder.
 
-## Mock 服务精简
+## Remove Mock Service
 
 If you don’t need the `Mock` service, you can delete the `apps/backend-mock` folder. Also, remove the `VITE_NITRO_MOCK` variable from the `.env.development` file in your application.
